Move Orch2 to orch.h
